Flask-implementation/requirements.txt
flask
flask-wtf
wtforms
email-validator
flask-markdown
markdown
flask-login
flask-sqlalchemy
werkzeug
datetime
mkdocs